fix(telegram): drop dead media-root test seams and the harness split that cycled
The dispatch harness kept mocking `getAgentScopedMediaLocalRoots` after
`bot-message-dispatch.runtime.ts` stopped exporting it, which pushed the harness
past the 700-line budget and forced a split whose re-export created a runtime
import cycle (`check:guards`, `check:architecture`).

Delete the stale mock instead: the harness fits the budget again, the split file
is gone, and the cycle with it. Also drop the unused
`getAgentScopedMediaLocalRoots` / `recordInboundSessionMetaSafe` re-exports this
branch added to `bot-native-commands.runtime.ts`.
