+++ extensions/telegram/src/bot-message-dispatch.test-harness.ts
@@ -115,9 +115,6 @@ const resolveDefaultModelForAgentHoisted = vi.hoisted(() =>
   vi.fn(() => ({ provider: "openai", model: "gpt-test" })),
 );
 const resolveHumanDelayConfigHoisted = vi.hoisted(() => vi.fn());
-const getAgentScopedMediaLocalRootsHoisted = vi.hoisted(() =>
-  vi.fn((_cfg: unknown, agentId: string) => [`/tmp/.openclaw/workspace-${agentId}`]),
-);
 const resolveAgentScopedOutboundMediaAccessHoisted = vi.hoisted(() =>
   vi.fn((params: { agentId?: string }) => ({
     localRoots: [`/tmp/.openclaw/workspace-${params.agentId ?? "default"}`],
@@ -165,7 +162,6 @@ const modelSupportsVision = modelSupportsVisionHoisted;
 const resolveAgentDir = resolveAgentDirHoisted;
 const resolveDefaultModelForAgent = resolveDefaultModelForAgentHoisted;
 export const resolveHumanDelayConfig = resolveHumanDelayConfigHoisted;
-const getAgentScopedMediaLocalRoots = getAgentScopedMediaLocalRootsHoisted;
 export const resolveAgentScopedOutboundMediaAccess = resolveAgentScopedOutboundMediaAccessHoisted;
 const resolveChunkMode = resolveChunkModeHoisted;
 export const resolveMarkdownTableMode = resolveMarkdownTableModeHoisted;
@@ -301,7 +297,6 @@ vi.mock("./send.js", async () => ({
 vi.mock("./bot-message-dispatch.runtime.js", () => ({
   generateTopicLabel: generateTopicLabelHoisted,
   getSessionEntry: getSessionEntryHoisted,
-  getAgentScopedMediaLocalRoots: getAgentScopedMediaLocalRootsHoisted,
   resolveAgentScopedOutboundMediaAccess: resolveAgentScopedOutboundMediaAccessHoisted,
   resolveAutoTopicLabelConfig: resolveAutoTopicLabelConfigRuntime,
   resolveChunkMode: resolveChunkModeHoisted,
@@ -417,7 +412,6 @@ function resetTelegramDispatchTestState() {
   loadSessionStore.mockReset();
   resolveStorePath.mockReset();
   generateTopicLabel.mockReset();
-  getAgentScopedMediaLocalRoots.mockClear();
   resolveAgentScopedOutboundMediaAccess.mockClear();
   resolveChunkMode.mockClear();
   resolveMarkdownTableMode.mockClear();
@@ -675,12 +669,74 @@ export function createRuntime(): Parameters<typeof dispatchTelegramMessage>[0]["
   };
 }
 
-export {
-  createReasoningDefaultContext,
-  createReasoningForumTopicContext,
-  createReasoningStreamContext,
-  dispatchWithContext,
-} from "./bot-message-dispatch-contexts.test-harness.js";
+export async function dispatchWithContext(params: {
+  context: TelegramMessageContext;
+  cfg?: Parameters<typeof dispatchTelegramMessage>[0]["cfg"];
+  telegramCfg?: Parameters<typeof dispatchTelegramMessage>[0]["telegramCfg"];
+  streamMode?: Parameters<typeof dispatchTelegramMessage>[0]["streamMode"];
+  telegramDeps?: TelegramBotDeps;
+  bot?: Bot;
+  replyToMode?: Parameters<typeof dispatchTelegramMessage>[0]["replyToMode"];
+  retryDispatchErrors?: boolean;
+  suppressFailureFallback?: boolean;
+  textLimit?: number;
+  turnAdoptionLifecycle?: Parameters<typeof dispatchTelegramMessage>[0]["turnAdoptionLifecycle"];
+  runtime?: Parameters<typeof dispatchTelegramMessage>[0]["runtime"];
+  opts?: Parameters<typeof dispatchTelegramMessage>[0]["opts"];
+}) {
+  const bot = params.bot ?? createBot();
+  return await dispatchTelegramMessage({
+    context: params.context,
+    bot,
+    cfg: params.cfg ?? {},
+    runtime: params.runtime ?? createRuntime(),
+    replyToMode: params.replyToMode ?? "first",
+    streamMode: params.streamMode ?? "partial",
+    textLimit: params.textLimit ?? 4096,
+    telegramCfg: params.telegramCfg ?? {},
+    telegramDeps: params.telegramDeps ?? telegramDepsForTest,
+    opts: params.opts ?? { token: "token" },
+    retryDispatchErrors: params.retryDispatchErrors,
+    suppressFailureFallback: params.suppressFailureFallback,
+    turnAdoptionLifecycle: params.turnAdoptionLifecycle,
+  });
+}
+
+export function createReasoningStreamContext(): TelegramMessageContext {
+  loadSessionStore.mockReturnValue({
+    s1: { reasoningLevel: "stream" },
+  });
+  return createContext({
+    ctxPayload: { SessionKey: "s1" } as unknown as TelegramMessageContext["ctxPayload"],
+  });
+}
+
+export function createReasoningDefaultContext(): TelegramMessageContext {
+  loadSessionStore.mockReturnValue({
+    s1: {},
+  });
+  return createContext({
+    ctxPayload: { SessionKey: "s1" } as unknown as TelegramMessageContext["ctxPayload"],
+    route: { agentId: "ops" } as unknown as TelegramMessageContext["route"],
+  });
+}
+
+export function createReasoningForumTopicContext(): TelegramMessageContext {
+  loadSessionStore.mockReturnValue({
+    s1: { reasoningLevel: "stream" },
+  });
+  return createContext({
+    ctxPayload: { SessionKey: "s1" } as unknown as TelegramMessageContext["ctxPayload"],
+    msg: {
+      chat: { id: -100123, type: "supergroup", is_forum: true },
+      message_id: 456,
+      message_thread_id: 88,
+    } as unknown as TelegramMessageContext["msg"],
+    chatId: -100123,
+    isGroup: true,
+    threadSpec: { id: 88, scope: "forum" },
+  });
+}
 
 export function describeTelegramDispatch(name: string, registerTests: () => void): void {
   describe(name, () => {
