+++ extensions/telegram/src/bot-native-commands.runtime.ts
@@ -2,12 +2,7 @@ import type { OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
 // Telegram plugin module implements bot native commands behavior.
 import { resolveAgentScopedOutboundMediaAccess } from "openclaw/plugin-sdk/media-local-roots";
 
-export {
-  ensureConfiguredBindingRouteReady,
-  recordInboundSessionMetaSafe,
-} from "openclaw/plugin-sdk/conversation-runtime";
-export { getAgentScopedMediaLocalRoots } from "openclaw/plugin-sdk/media-local-roots";
-export { resolveAgentScopedOutboundMediaAccess };
+export { ensureConfiguredBindingRouteReady } from "openclaw/plugin-sdk/conversation-runtime";
 export {
   finalizeInboundContext,
   resolveChunkMode,
@@ -16,10 +11,10 @@ export { resolveThreadSessionKeys } from "openclaw/plugin-sdk/routing";
 export { getSessionEntry } from "openclaw/plugin-sdk/session-store-runtime";
 
 /**
- * Resolves policy-gated local media roots for a native command reply delivery.
- * Configured `agents.defaults.mediaLocalRoots` are honored only when the
- * sender/group host-media-read policy allows; the generic helper never grants
- * them as an ambient authorization.
+ * Owns the Telegram-side inputs to the shared outbound media policy: pins the
+ * provider, derives the group id, and forwards requester identity. Configured
+ * `agents.defaults.mediaLocalRoots` apply only when sender/group host-read
+ * policy allows — the generic root helper never grants them ambiently.
  */
 export function resolveNativeCommandOutboundMediaRoots(params: {
   cfg: OpenClawConfig;
